Refactor access mode handling in filesystem headers
- Updated `dos_faccessat` and `native_faccessat` functions to correctly use the `access_how` parameter instead of `dos_at_flags` for mode handling.
- Adjusted conditional checks in `nt_at.h` and `win32_9xa_at.h` to ensure proper evaluation of `symlink_nofollow` flag.
- Improved clarity and consistency in the handling of file access modes across different platform implementations.

+++ include/fast_io_hosted/filesystem/win32_9xa_at.h
@@ -56,20 +56,19 @@ inline void win32_9xa_unlinkat_impl(::fast_io::win32_9xa_dir_handle const &dirhd
 {
 	auto file_or_path{concat_tlc_win32_9xa_path_uncheck_whether_exist(dirhd, path_c_str, path_size)};
 
-	switch (flags)
+	if ((flags & ::fast_io::win32_9xa_at_flags::removedir) == ::fast_io::win32_9xa_at_flags::removedir)
 	{
-	case ::fast_io::win32_9xa_at_flags::removedir:
 		if (!::fast_io::win32::RemoveDirectoryA(reinterpret_cast<char const *>(file_or_path.c_str()))) [[unlikely]]
 		{
 			throw_win32_error();
 		}
-		break;
-	default:
+	}
+	else
+	{
 		if (!::fast_io::win32::DeleteFileA(reinterpret_cast<char const *>(file_or_path.c_str()))) [[unlikely]]
 		{
 			throw_win32_error();
 		}
-		break;
 	}
 }
 
@@ -218,7 +217,7 @@ inline constexpr auto calculate_win32_9xa_readonly_open_mode(bool write_attribut
 inline posix_file_status win32_9xa_fstatat_impl(::fast_io::win32_9xa_dir_handle const &dirhd, char8_t const *path_c_str, ::std::size_t path_size, win32_9xa_at_flags flags)
 {
 	auto path{concat_tlc_win32_9xa_path_uncheck_whether_exist(dirhd, path_c_str, path_size)};
-	auto md{calculate_win32_9xa_readonly_open_mode(false, (flags & win32_9xa_at_flags::symlink_nofollow) != win32_9xa_at_flags::symlink_nofollow)};
+	auto md{calculate_win32_9xa_readonly_open_mode(false, (flags & win32_9xa_at_flags::symlink_nofollow) == win32_9xa_at_flags::symlink_nofollow)};
 	::fast_io::win32_file_9xa f{::fast_io::details::win32_family_create_file_internal_impl<win32_family::ansi_9x>(reinterpret_cast<char const *>(path.c_str()), md)};
 
 	return ::fast_io::win32::details::win32_status_impl(f.native_handle());
@@ -228,7 +227,7 @@ inline void win32_9xa_utimensat_impl(::fast_io::win32_9xa_dir_handle const &dirh
 									 unix_timestamp_option last_access_time, unix_timestamp_option last_modification_time, win32_9xa_at_flags flags)
 {
 	auto path{concat_tlc_win32_9xa_path_uncheck_whether_exist(dirhd, path_c_str, path_size)};
-	auto md{calculate_win32_9xa_readonly_open_mode(true, (flags & win32_9xa_at_flags::symlink_nofollow) != win32_9xa_at_flags::symlink_nofollow)};
+	auto md{calculate_win32_9xa_readonly_open_mode(true, (flags & win32_9xa_at_flags::symlink_nofollow) == win32_9xa_at_flags::symlink_nofollow)};
 	::fast_io::win32_file_9xa f{::fast_io::details::win32_family_create_file_internal_impl<win32_family::ansi_9x>(reinterpret_cast<char const *>(path.c_str()), md)};
 
 	::fast_io::win32::filetime ftm;
